+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 ### New in 0.39 (not released yet)
 
-* _Nothing yet_
+* New: `Command<,,>` now inherits from `ValueObject` 
 
 ### New in 0.38.2454 (released 2016-12-02)