+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
-### New in 0.38 (not released yet)
+### New in 0.39 (not released yet)
 
-* _Nothing yet_
+* New: `Command<,,>` now inherits from `ValueObject` 
+
+### New in 0.38.2454 (released 2016-12-02)
+
+* Fix: Single aggregate read models can now be re-populated again
 
 ### New in 0.37.2424 (released 2016-11-08)